internal/pretty: moves the package from under `cmd` folder.
