+++ cmd/gohandlers/commands/helpers/internal/imports/imports.go
@@ -4,7 +4,7 @@ import (
 	"go/ast"
 	"go/token"
 
-	"go.ufukty.com/gohandlers/cmd/gohandlers/internal/pretty/sort"
+	"go.ufukty.com/gohandlers/internal/pretty/sort"
 	"go.ufukty.com/gohandlers/pkg/inspects"
 )
